requirements.txt
google-generativeai
streamlit
pypdf
gtts

